% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{$.hdd}
\alias{$.hdd}
\title{Extracts a single variable from a HDD object}
\usage{
\method{$}{hdd}(x, name)
}
\arguments{
\item{x}{A hdd object.}

\item{name}{The variable name to be extracted.Note that there is an automatic protection for not trying to import data that would not fit into memory. The extraction cap is set with the function \code{\link[hdd]{setHdd_extract.cap}}.}
}
\value{
It returns a vector.
}
\description{
This method extracts a single variable from a hard drive data set (HDD). There is an automatic protection to avoid extracting too large data into memory. The bound is set by the function \code{\link[hdd]{setHdd_extract.cap}}.
}
\examples{

\dontrun{
# your data set is in the hard drive, in hdd format already.
# Say this data is made of two variables: x and id.
# 'x' is integer and 'id' is of type character.
data_hdd = hdd("path/big_data")

# The two variables, if loaded into memory, would be of size:
# - x :  800MB
# - id: 5000MB

# There is a protection for not loading variable that are too large.
# By default, the cap is set at 1000MB, you can change it
# with setHdd_extract.cap(new_cap).

# By default, the following would run:
x = data_hdd$x

# But this would raise an error:
id = data_hdd$id

# To make it would you need at least to set a 5GB cap:
setHdd_extract.cap(6000)
id = data_hdd$id
# Now it would work (provided you have enough memory!!!)

}

}
