% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{guess_col_types}
\alias{guess_col_types}
\title{Guesses the columns types of a file}
\usage{
guess_col_types(dt_or_path, col_names, n = 10000)
}
\arguments{
\item{dt_or_path}{Either a data frame or a path.}

\item{col_names}{Optional: the vector of names of the columns, if not contained in the file. Must match the number of columns in the file.}

\item{n}{Number of observations used to make the guess. By default, \code{n = 100000}.}
}
\value{
It returns a \code{cols} object a la \code{readr.}
}
\description{
This function is a facility to guess the column types of a text document. It returns columns formatted a la readr.
}
\examples{

\dontrun{

# Find the columns of a text file
guess_cols("path/my_data.txt")

}


}
