% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{guess_delim}
\alias{guess_delim}
\title{Guesses the delimiter of a text file}
\usage{
guess_delim(path)
}
\arguments{
\item{path}{The path to a text file containing a rectangular data set.}
}
\value{
It returns a character string of length 1: the delimiter.
}
\description{
This function uses \code{\link[data.table]{fread}} to guess the delimiter of a text file.
}
\examples{

\dontrun{

# Say you have data in the text file "path/data.txt".
# To guess the delimiter:
guess_delim("path/data.txt")

}

}
