% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{hdd}
\alias{hdd}
\title{Hard drive data set}
\usage{
hdd(dir)
}
\arguments{
\item{dir}{The directory where the hard drive data set is.}
}
\description{
This function connects to a hard drive data set (hdd). You can access the hard drive data in a similar way as a \code{data.table}.
}
\examples{

\dontrun{
# your data set is in the hard drive, in hdd format already
data_hdd = hdd("path/my_big_data")
summary(data_hdd)

}


}
