% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{hdd_setkey}
\alias{hdd_setkey}
\title{Sorts HDD objects}
\usage{
hdd_setkey(x, key, newfile, chunkMB = 500, replace = FALSE,
  verbose = 1)
}
\arguments{
\item{x}{A hdd file.}

\item{key}{A character vector of the keys.}

\item{newfile}{Destination of the result, i.e., a destination folder that will receive the HDD data.}

\item{chunkMB}{The size of chunks used to sort the data. Default is 500MB. The bigger this number the faster the sorting is (depends on your memory available though).}

\item{replace}{Default is \code{FALSE}: if the destination folder already contains data, whether to replace it.}
}
\description{
This function sets a key to a HDD file. It creates a copy of the HDD file sorted by the key. Note that the sorting process is very time consuming.
}
\examples{

\dontrun{
# your data set is in the hard drive, in hdd format already.
# Say this data is made of two variables: x and id.
data_hdd = hdd("path/big_data")

# We want to create a new HDD file, sorted by id:
hdd_setkey(data_hdd, "id", "path/big_data_sorted")

}

}
