% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{names.hdd}
\alias{names.hdd}
\title{Variables names of a HDD object}
\usage{
\method{names}{hdd}(x)
}
\arguments{
\item{x}{A \code{HDD} object.}
}
\value{
A character vector.
}
\description{
Gets the variable names of a hard drive data set (HDD).
}
\examples{

\dontrun{
# your data set is in the hard drive, in hdd format already
data_hdd = hdd("path/my_big_data")
names(data_hdd)

}

}
\author{
Laurent Berge
}
