% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{readfst}
\alias{readfst}
\title{Read fst or HDD files as DT}
\usage{
readfst(path, columns = NULL, from = 1, to = NULL, confirm = FALSE)
}
\arguments{
\item{path}{Path to \code{fst} file -- or path to \code{hdd} data. For hdd files, there is a}

\item{columns}{Column names to read. The default is to read all columns. Ignored for \code{hdd} files.}

\item{from}{Read data starting from this row number. Ignored for \code{hdd} files.}

\item{to}{Read data up until this row number. The default is to read to the last row of the stored data set. Ignored for \code{hdd} files.}

\item{confirm}{If the hdd file is larger than the variable \code{getHdd_extract.cap()}, then by default an error is raised. To anyway read the data, use \code{confirm = TRUE}. You can set the data cap with the function \code{\link[hdd]{setHdd_extract.cap}}, the default being 1GB.}
}
\description{
This is the function \code{\link[fst]{read_fst}} but with automatic conversion to data.table. It also allows to read \code{hdd} data.
}
\examples{

\dontrun{
base = readfst("path.fst")
# is exactly equivalent to:
base = read_fst("path.fst", as.data.table = TRUE)

# reading the full data from a HDD file
base = readfst("hdd_path")
# is equivalent to:
base = hdd("hdd_path")[]

}




}
\author{
Laurent Berge
}
