% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{setkeyhdd}
\alias{setkeyhdd}
\title{Sorts HDD objects}
\usage{
setkeyhdd(x, key, newfile, chunkMB = 500)
}
\arguments{
\item{x}{A hdd file.}

\item{key}{A character vector of the keys.}

\item{newfile}{Where to save the sorted file.}

\item{chunkMB}{The size of chunks used to sort the data. Default is 500MB. The bigger this number the faster the sorting is (depends on your memory available though).}
}
\description{
This function sets a key to a HDD file. It creates a copy of the HDD file sorted by the key. Note that the sorting process is very time consuming.
}
\examples{

\dontrun{
# your data set is in the hard drive, in hdd format already.
# Say this data is made of two variables: x and id.
data_hdd = hdd("path/big_data")

# We want to create a new HDD file, sorted by id:
setkeyhdd(data_hdd, "id", "path/big_data_sorted")

}

}
