% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{txt2hdd}
\alias{txt2hdd}
\title{Transforms text data into a HDD file}
\usage{
txt2hdd(path, dirDest, chunkMB, col_names, col_types, nb_skip, delim,
  preprocessfun, replace = FALSE, verbose = 1, ...)
}
\arguments{
\item{path}{Path where the data is.}

\item{dirDest}{The destination directory, where the new HDD data should be saved.}

\item{chunkMB}{The chunk sizes in MB, no default.}

\item{col_names}{The column names, by default is uses the ones of the dataset. If the dataset lacks column names, you must provide them.}

\item{col_types}{The column types, in the \code{readr} fashion. You can use \code{\link{guess_col_types}} to find them.}

\item{nb_skip}{Number of lines to skip.}

\item{delim}{The delimiter. By default the function tries to find the delimiter, but sometimes it fails.}

\item{preprocessfun}{A function that is applied to the data before saving. Default is missing. Note that if a function is provided, it MUST return a data.frame, anything other than data.frame is ignored.}

\item{replace}{If the destination directory already exists, you need to set the argument \code{replace=TRUE} to overwrite all the HDD files in it.}

\item{verbose}{Integer. If verbose > 0, then the evolution of the importing process is reported.}

\item{...}{Other arguments to be passed to \code{\link[readr]{read_delim_chunked}}, \code{quote = ""} can be interesting sometimes.}
}
\description{
Imports text data and saves it into a HDD file. It uses \code{\link[readr]{read_delim_chunked}} to extract the data. It also allows to preprocess the data.
}
\examples{

\dontrun{

# Example in the context of publication data.
# Assume we have a large text file containing the following variables:
# - publication_id: unique publication identifier
# - abstract: the abstract of the publicaiton
# - ...: many other variables
#
# We want to import the data, but keep only the observations for which
# the abstract is written in English.
# We have created beforehand the function 'check_english' which checks if
# a text is in English language.

# To import the original data, but only with English abstract,
# you can do:
fun = function(x){
  x[check_english(abstract)]
}

txt2hdd("abstracts.txt", "path/pub_english", chunkMB = 1000, preprocessfun = fun)
}


}
\author{
Laurent Berge
}
