% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{dim.hdd}
\alias{dim.hdd}
\title{Dimension of a HDD object}
\usage{
\method{dim}{hdd}(x)
}
\arguments{
\item{x}{A \code{HDD} object.}
}
\value{
It returns a vector of length 2 containing the number of rows and the columns.
}
\description{
Gets the dimension of a hard drive data set (HDD).
}
\examples{

\dontrun{
# your data set is in the hard drive, in hdd format already
data_hdd = hdd("path/my_big_data")
dim(data_hdd)

}

}
\author{
Laurent Berge
}
