% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{hdd_merge}
\alias{hdd_merge}
\title{Merges data to a HDD file}
\usage{
hdd_merge(x, y, newfile, all = FALSE, all.x = all, all.y = all,
  replace = FALSE, verbose)
}
\arguments{
\item{x}{A hdd file.}

\item{y}{A data set either a data.frame of a HDD object.}

\item{newfile}{Destination of the result, i.e., a destination folder that will receive the HDD data.}

\item{all}{Default is \code{FALSE}.}

\item{all.x}{Default is \code{all}.}

\item{all.y}{Default is \code{all}.}

\item{replace}{Default is \code{FALSE}: if the destination folder already contains data, whether to replace it.}

\item{verbose}{Numeric. Whether information on the advancement should be displayed. If equal to 0, nothing is displayed. By default it is equal to 1 if the size of \code{x} is greater than 1GB.}
}
\description{
This function merges in-memory/HDD data to a HDD file.
}
\examples{

\dontrun{

y = data.frame(xx = 1:3, Species = c("virginica", "setosa", "versicolor"))
iris_hdd = hdd("iris_hdd")

hdd_merge(iris_hdd, y, "iris_merged_with_y")

}

}
\author{
Laurent Berge
}
