% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{peek}
\alias{peek}
\title{Peek into a text file}
\usage{
peek(path, onlyLines = FALSE, n, view = TRUE)
}
\arguments{
\item{path}{Path linking to the text file.}

\item{onlyLines}{Default is \code{FALSE}. If \code{TRUE}, then the first \code{n} lines are directly displayed without formatting.}

\item{n}{Integer. The number of lines to extract from the file. Default is 100 or 5 if \code{onlyLine = TRUE}.}

\item{view}{Logical, default it \code{TRUE}: whether the data should be displayed on the viewer. Only when \code{onlyLines = FALSE}.}
}
\value{
Returns the data invisibly.
}
\description{
This function looks at the first elements of a file, format it into a data frame and displays it. It can also just show the first lines of the file without formatting into a DF.
}
\examples{

\dontrun{

# Let's have a look at the first observations from "path/data.txt")
peek("path/data.txt")

# Sometimes, it can be useful to look at the unformatted lines:
peek("path/data.txt", onlyLines = TRUE)

}

}
\author{
Laurent Berge
}
