% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.hdd}
\alias{print.hdd}
\title{Print method for hdd objects}
\usage{
\method{print}{hdd}(x, ...)
}
\arguments{
\item{x}{A hdd object.}

\item{...}{Not currently used.}
}
\value{
Nothing is returned.
}
\description{
This functions displays the first and last lines of a hard drive data set (hdd).
}
\examples{

\dontrun{
# your data set is in the hard drive, in hdd format already
data_hdd = hdd("path/my_big_data")
print(data_hdd)

}

}
