% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setHdd_extract.cap}
\alias{setHdd_extract.cap}
\alias{getHdd_extract.cap}
\title{Sets/gets the size cap when extracting hdd data}
\usage{
setHdd_extract.cap(sizeMB = 1000)

getHdd_extract.cap()
}
\arguments{
\item{sizeMB}{Size cap in MB. Default to 1000.}
}
\value{
The size cap, a numeric scalar.
}
\description{
Sets/gets the default size cap when extracting \code{hdd} data. If the size exceeds the cap, then an error is raised, which can be bypassed by using the argument \code{confirm}.
}
