% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{summary.hdd}
\alias{summary.hdd}
\title{Summary information for hdd objects}
\usage{
\method{summary}{hdd}(object, ...)
}
\arguments{
\item{object}{A hdd object.}

\item{...}{Not currently used.}
}
\description{
Provides summary information -- i.e. dimension, size on disk, path, number of slices -- of hard drive data sets (hdd).
}
\examples{

\dontrun{
# your data set is in the hard drive, in hdd format already
data_hdd = hdd("path/my_big_data")
summary(data_hdd)

}

}
