% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write_hdd}
\alias{write_hdd}
\title{Saves or appends a data set into a HDD file}
\usage{
write_hdd(x, dir, chunkMB = Inf, compress = 50, add = FALSE,
  replace = FALSE, showWarning, ...)
}
\arguments{
\item{x}{A data set.}

\item{dir}{The HDD repository, i.e. the directory where the HDD data is.}

\item{chunkMB}{If the data has to be split in several files of \code{chunkMB} sizes. Default is \code{Inf}.}

\item{compress}{Compression rate to be applied by \code{\link[fst]{write_fst}}. Default is 50.}

\item{add}{Should the file be added to the existing repository? Default is \code{FALSE}.}

\item{replace}{If \code{add = FALSE}, should any existing document be replaced? Default is \code{FALSE}.}

\item{showWarning}{If the data \code{x} has no observation, then a warning is raised if \code{showWarning = TRUE}. By default, it occurs only if \code{write_hdd} is NOT called within a function.}

\item{...}{Not currently used.}
}
\description{
This function saves in-memory/HDD data sets into HDD repositories. Useful to append several data sets.
}
\examples{

\dontrun{

# Say you have a data set split into 25 different files.
# You want to save them all in a single HDD file:
write_hdd(data_01, "path/big_data")
write_hdd(data_02, "path/big_data", add = TRUE)
# etc...
write_hdd(data_25, "path/big_data", add = TRUE)

# Now say you want to reformat a HDD file
# to increase the chunk sizes:
write_hdd("path/big_data", "path/big_data_smallChunk", chunkMB = 1000)

}

}
\author{
Laurent Berge
}
